return form url util
