+++ @xen-orchestra/web/src/modules/backup/components/repository/list/panel/BackupRepositorySidePanel.vue
@@ -3,11 +3,14 @@
     <template v-if="br" #default>
       <BackupRepositoryInfosCard :br />
       <BackupRepositorySpaceAndSpeedCard :br />
-      <BackupRepositoryLocalCard v-if="brInfo?.type === 'file'" :file="brInfo" />
-      <BackupRepositoryNfsCard v-else-if="brInfo?.type === 'nfs'" :nfs="brInfo" :options="br.options" />
-      <BackupRepositorySmbCard v-else-if="brInfo?.type === 'smb'" :smb="brInfo" :options="br.options" />
-      <BackupRepositoryS3Card v-else-if="brInfo?.type === 's3'" :s3="brInfo" />
-      <BackupRepositoryAzureCard v-else-if="brInfo?.type === 'azure' || brInfo?.type === 'azurite'" :azure="brInfo" />
+      <BackupRepositoryLocalCard v-if="parsedBrUrl?.type === 'file'" :file="parsedBrUrl" />
+      <BackupRepositoryNfsCard v-else-if="parsedBrUrl?.type === 'nfs'" :nfs="parsedBrUrl" :options="br.options" />
+      <BackupRepositorySmbCard v-else-if="parsedBrUrl?.type === 'smb'" :smb="parsedBrUrl" :options="br.options" />
+      <BackupRepositoryS3Card v-else-if="parsedBrUrl?.type === 's3'" :s3="parsedBrUrl" />
+      <BackupRepositoryAzureCard
+        v-else-if="parsedBrUrl?.type === 'azure' || parsedBrUrl?.type === 'azurite'"
+        :azure="parsedBrUrl"
+      />
     </template>
   </VtsSidePanel>
 </template>
@@ -21,7 +24,7 @@ import BackupRepositoryS3Card from '@/modules/backup/components/repository/list/
 import BackupRepositorySmbCard from '@/modules/backup/components/repository/list/panel/cards/BackupRepositorySmbCard.vue'
 import BackupRepositorySpaceAndSpeedCard from '@/modules/backup/components/repository/list/panel/cards/BackupRepositorySpaceAndSpeedCard.vue'
 import { type FrontXoBackupRepository } from '@/modules/backup/remote-resources/use-xo-backup-repository-collection.ts'
-import { parseBackupRepositoryUrl } from '@/modules/backup/utils/xo-backup-repository-url.util.ts'
+import { parseBackupRepositoryUrl } from '@/modules/backup/utils/xo-parse-backup-repository-url.util.ts'
 import VtsSidePanel from '@core/components/panel/VtsSidePanel.vue'
 import { computed } from 'vue'
 
@@ -33,5 +36,5 @@ const emit = defineEmits<{
   close: []
 }>()
 
-const brInfo = computed(() => (br === undefined ? undefined : parseBackupRepositoryUrl(br.url)))
+const parsedBrUrl = computed(() => (br === undefined ? undefined : parseBackupRepositoryUrl(br.url)))
 </script>
