+++ @xen-orchestra/web/src/modules/backup/components/repository/list/panel/cards/BackupRepositoryLocalCard.vue
@@ -14,7 +14,7 @@
 </template>
 
 <script lang="ts" setup>
-import { type FileBackupRepositoryInfo } from '@/modules/backup/utils/xo-backup-repository-url.util.ts'
+import { type FileBackupRepositoryInfo } from '@/modules/backup/utils/xo-parse-backup-repository-url.util.ts'
 import VtsCardRowKeyValue from '@core/components/card/VtsCardRowKeyValue.vue'
 import VtsCopyButton from '@core/components/copy-button/VtsCopyButton.vue'
 import UiCardTitle from '@core/components/ui/card-title/UiCardTitle.vue'
